requirements.txt
pyudev~=0.23.2
ipython~=7.20.0
matplotlib~=3.3.4
numpy~=1.21.5
scipy~=1.7.3
PyAudio~=0.2.11
Pillow~=8.4.0
websockets~=10.3
gTTS~=2.2.4
playsound~=1.3.0
wolframalpha~=5.0.0
pyttsx3~=2.90
wikipedia~=1.4.0
pyjokes~=0.6.0
feedparser~=6.0.10
requests~=2.25.1
twilio~=7.9.1
beautifulsoup4~=4.9.3
uvloop~=0.16.0
pyserial~=3.5
setuptools~=62.1.0
mutagen~=1.45.1
docutils~=0.18.1
lxml~=4.6.3
PyQt5~=5.15.2
olefile~=0.46
soupsieve~=2.2.1
html5lib~=1.1
cryptography~=3.3.2
pbr~=5.8.0
six~=1.16.0
pip~=22.1.1
distro~=1.5.0
colorlog~=6.6.0
tornado~=6.1
pyzmq~=20.0.0
pexpect~=4.8.0
gevent~=21.12.0
click~=8.1.2
mock~=4.0.3
urllib3~=1.26.5
parso~=0.8.1
jedi~=0.18.0
mypy~=0.950
tomli~=2.0.1
pytz~=2022.1
PyYAML~=5.3.1
greenlet~=1.1.2
psutil~=5.8.0
pyglet~=1.5.26
PyGObject~=3.38.0
PyJWT~=2.4.0
Pygments~=2.7.1
ipykernel~=5.4.3
traitlets~=5.0.5
decorator~=5.1.0
backcall~=0.2.0
pickleshare~=0.7.5
wcwidth~=0.1.9
certifi~=2020.6.20
chardet~=4.0.0
pycurl~=7.43.0.6
python-dateutil~=2.8.1
graphviz~=0.20
idna~=2.10
asttokens~=2.0.5
ply~=3.11
websocket~=0.2.1
websocket-client~=1.3.2
sgmllib3k~=1.0.0
cycler~=0.10.0
pyparsing~=2.4.7
kiwisolver~=1.3.1
xmltodict~=0.13.0
board~=1.0
imageio~=2.13.5